Create a Map server that will be responsible for answering client requests
Unify the components structure so that each component has:
  - An interface (Participant, Coordinator, MapNode)
  - A base implementation (BaseParticipant, BaseCoordinator, BaseMapNode)
  - A controller (ParticipantController, CoordinatorController, MapNodeController)

While the base implementation codifies the "business logic" of the component and handles all the state, communication with IO (such as sockets) is done through the controller.

Each component that implements the interface `Controllable` must provide the methods `setController` and `getController`.

Each controller generally receives the component in it's constructor and calls `setController` on said component, passing a reference to itself as an argument. Then, when initializing, it initializes the component first, and then initializes itself.

The `MapNode` uses a `Participant` to handle the transactions, and is responsible for storing the Map state, and responding to requests from the Map clients.

The DistributedTransactions already stores the transactions in the logs, but the loading phase in case of a restart are still a WIP.
Also once a transaction is written to a log, it is never discarded, even after it is no longer in use. That has yet to be implemented.